refactor(rath): split ui components of LaTiaoConsole
